+++ library/data-model/src/utils.ts
@@ -29,6 +29,17 @@ export const slugify = (str: string) => {
 };
 
 /**
+ * Formats file size in human-readable format
+ */
+export function formatFileSize(bytes: number): string {
+  if (bytes === 0) return '0 Bytes';
+  const k = 1024;
+  const sizes = ['Bytes', 'KB', 'MB', 'GB'];
+  const i = Math.floor(Math.log(bytes) / Math.log(k));
+  return Math.round((bytes / Math.pow(k, i)) * 100) / 100 + ' ' + sizes[i];
+}
+
+/**
  * Finds the difference between sets A and B, i.e. the elements which are in A,
  * but not in B.
  */
